run-meecrowave{ => -deltaspike}: rename module and script
Ahead of adding a run-meecrowave-jpa module and bin/meecrowave-jpa-run.sh script.
