--- bin/meecrowave-run.sh
@@ -1,6 +1,0 @@
-#!/usr/bin/env bash
-
-set -e
-
-mvn install -pl run-meecrowave -am
-mvn meecrowave:run -pl run-meecrowave
