+++ run-meecrowave-deltaspike/src/main/resources/META-INF/beans.xml
